DFG should be able to compute dominators https://bugs.webkit.org/show_bug.cgi?id=85269
Unreviewed trivial 32 bit buildfix after r117861.

* dfg/DFGGraph.cpp:
(JSC::DFG::Graph::dump):

Canonical link: https://commits.webkit.org/104758@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117905 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/dfg/DFGGraph.cpp
@@ -288,14 +288,14 @@ void Graph::dump()
             for (size_t i = 0; i < m_blocks.size(); ++i) {
                 if (!m_dominators.dominates(i, b))
                     continue;
-                dataLog(" #%lu", i);
+                dataLog(" #%lu", static_cast<unsigned long>(i));
             }
             dataLog("\n");
             dataLog("  Dominates:");
             for (size_t i = 0; i < m_blocks.size(); ++i) {
                 if (!m_dominators.dominates(b, i))
                     continue;
-                dataLog(" #%lu", i);
+                dataLog(" #%lu", static_cast<unsigned long>(i));
             }
             dataLog("\n");
         }
